requirements.txt
torch==1.9.0
torchvision==0.10.0
Pillow==8.4.0
pandas==1.3.4
transformers==4.13.0